jinx-correct: Do not restore point position
When spell checking the whole buffer one sometimes wants to make a cursory edit
at the location of the misspelling. Therefore it is a better behavior to not
restore the point position. Instead push a mark which allows you to go back to
where you came from before spell-checking.

See discussion in #9.