Add exclude faces for git-commit-mode
Git-commit mode contains two parts of text, the first part is the
actual content mixed with several words that should be excluded that
use git-commit-mode faces.
The second part is the comments as part of the markdown-mode used,
the other faces that should be excluded in there can be ignored
because the markdown-comment-face excluding them already.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>