Mention deps rpm distros with pkgconf and openSUSE package
In rpm based distributions dependencies can be installed using
pkgconfig symbols instead of plain package names, the benefit is that
this will use the pkgconfig name which is also used in the jinx itself
and the name of package doesn't matter anymore.

Also mention openSUSE Tumbleweed package similar to the Guix package
which is already mentioned.

Signed-off-by: Björn Bidar <bjorn.bidar@thaodan.de>