Stop using enchant_dict_get_extra_word_characters
Enchant implements this function for the Hunspell backend only, such that many
users do not profit from it. See if we get by well without it, relying on the
default syntax table only. Test this change for a while to see how well it
works. See also https://github.com/rrthomas/enchant/issues/244.